``` refactor: inline data fetching logic and remove useTableData hook from founder table
- Removed useTableData custom hook and moved query logic directly into FounderViewV2 component
- Added useMemo for derived data (founders, totalPages, totalItems) to prevent unnecessary recalculations
- Imported useQuery and ListFoundersResponse types directly in FounderViewV2
- Removed useTableData export from table-v2 index
- Deleted use-table-data.ts file as logic is now colocated with component
- Simplifie
